requirements.txt
streamlit
opencv-python
numpy
scikit-learn
joblib
tensorflow
Pillow
matplotlib
seaborn
xgboost
